requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
cffi==1.17.1
channels==3.0.4
constantly==23.10.4
cryptography==44.0.0
daphne==3.0.2
Django==4.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==10.4.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
twisted==24.11.0
txaio==23.1.1
typing-extensions==4.12.2
tzdata==2024.2
zope.interface==7.2
